+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Unreleased
 - Add function `name_jsstring` to `SharedCreepProperties` trait, `Flag`, and `StructureSpawn`
 - Add function `raw_memory::segments_jsstring` to get the object as `JsHashMap<u8, JsString>`
 - Add `RoomTerrain::get_xy` which takes a `RoomXY` instead of a coordinate pair
-- Add `RoomCoordinate::is_edge` and `RoomXY::is_edge`
+- Add `RoomCoordinate::is_room_edge`, `RoomXY::is_room_edge`, and `Position::is_room_edge`
 - Add new extra constant `constants::extra::CONTROLLER_DOWNGRADE_PROGRESS_RATIO`, which is
   the percentage toward the previous level that a downgrading controller's progress is set to
 - Add new extra constants `constants::extra::MOVE_COST_PLAIN`, `MOVE_COST_ROAD`, `MOVE_COST_SWAMP`